{{cookiecutter.repo_name}}/requirements.txt
black
click
cryptography
flake8
isort
hdfs
pre-commit
PyMySQL
pytest
pytest-cov
python-dotenv
pytz
SQLAlchemy
tox